+++ tests/misc/programs/misc/src/lib.rs
@@ -190,4 +190,9 @@ pub mod misc {
     pub fn test_empty_seeds_constraint(ctx: Context<TestEmptySeedsConstraint>) -> ProgramResult {
         Ok(())
     }
+
+    pub fn test_init_if_needed(ctx: Context<TestInitIfNeeded>, data: u16) -> ProgramResult {
+        ctx.accounts.data.data = data;
+        Ok(())
+    }
 }